+++ packages/qiskit-qasm/lib/Parser.js
@@ -18,16 +18,15 @@ const utils = require('./utils');
 
 const dbg = utils.dbg(__filename);
 
-// const QasmError = require('./QasmError');
+const QasmError = require('./QasmError');
 
 // TODO: Do async?
 const bnf = fs.readFileSync(path.resolve(__dirname, 'grammar.jison'), 'utf8');
-let parser;
 
 class Parser {
   constructor(opts = {}) {
     dbg('Starting', opts);
-    parser = new jison.Parser(bnf);
+    this.parser = new jison.Parser(bnf);
 
     if (opts.core !== false) {
       // TODO: Parse all core libraries (when we have more)
@@ -35,23 +34,24 @@ class Parser {
         path.resolve(__dirname, '../core/qelib1.inc'),
         'utf8',
       );
-      this.qelibParsed = parser.parse(qelib1);
+      this.qelibParsed = this.parser.parse(qelib1);
     }
   }
 
   parse(circuit) {
     if (!circuit) {
-      throw new Error('Required param: circuit');
+      throw new TypeError('Required param: circuit');
     }
 
     let res;
 
     try {
-      res = parser.parse(circuit, this.qelibParsed);
+      res = this.parser.parse(circuit, this.qelibParsed);
     } catch (err) {
-      // TODO: Use our custom error
-      // throw new QasmError('')
-      throw err;
+      if (err instanceof QasmError) {
+        throw err;
+      }
+      throw new QasmError(err.message)
     }
 
     return res;